Renamed react-hooks package to use-subscription
