+++ packages/use-subscription/npm/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+if (process.env.NODE_ENV === 'production') {
+  module.exports = require('./cjs/use-subscription.production.min.js');
+} else {
+  module.exports = require('./cjs/use-subscription.development.js');
+}
