+++ packages/use-subscription/package.json
@@ -1,12 +1,12 @@
 {
   "private": true,
-  "name": "react-hooks",
+  "name": "use-subscription",
   "description": "Reusable hooks",
-  "version": "16.8.3",
+  "version": "16.8.6",
   "repository": {
     "type": "git",
     "url": "https://github.com/facebook/react.git",
-    "directory": "packages/react-hooks"
+    "directory": "packages/use-subscription"
   },
   "files": [
     "LICENSE",
@@ -16,7 +16,7 @@
     "cjs/"
   ],
   "peerDependencies": {
-    "react": "^16.3.0"
+    "react": "^16.8.0"
   },
   "devDependencies": {
     "rxjs": "^5.5.6"
